# static-contracts: improve optimizer's test for flat scs