This is one idea for fixing #628 .

I'm not happy about re-computing `recursive-kinds` in the optimizer,
but I don't see how to tell whether a `name/sc` is flat without it,
and I didn't see an easy way to share the optimizer's `recursive-kinds` with `instantiate.rkt`

- - -

Improve the static contract optimizer's ability to recognize flat contracts:

1. get the kind of all recursive, named contracts
2. any `name/sc` with `'flat` kind is a flat contract

Step 1 is repeated later in the `instantiate` pass.