@pnwamk 2017-10-14
Two questions/comments off the top of my head:

1. How does this affect compile time performance? Generally, and in code that spends a lot of time in contract generation (e.g. https://github.com/akuhlens/schml, and the schml examples in https://github.com/pnwamk/tr-performance)

2. More generally, I think we may just need the structs in our static-contract IRL to cache what kind of contract they are (i.e. flat, chaperone, impersonator) so we only compute this once not only for recursive name/sc's, but just in general for all static contracts. But perhaps this is a larger beast to tackle later (really we should just make a nice, clean DSL for defining static-contract structs like we have for Rep's)