@bennn 2017-10-16
1. Too much slower. Here's the current master branch vs. this branch:

```
acquire: 0.94 (i.e. slower, 18.19s (σ 0.42) to 19.41s (σ 0.21))
dungeon: 0.97 (i.e. slower, 10.08s (σ 0.08) to 10.36s (σ 0.59))
forth: 1.0 (i.e. faster, 5.6s (σ 0.05) to 5.6s (σ 0.05))
fsm: 1.0 (i.e. slower, 5.65s (σ 0.02) to 5.67s (σ 0.05))
math-flonum: 0.99 (i.e. slower, 4.53s (σ 0.02) to 4.56s (σ 0.04))
new-metrics: 1.06 (i.e. faster, 3.79s (σ 0.1) to 3.58s (σ 0.16))
old-metrics: 1.0 (i.e. faster, 2.87s (σ 0.04) to 2.86s (σ 0.02))
parser: 1.0 (i.e. faster, 2.43s (σ 0.02) to 2.42s (σ 0.01))
schml-interp-casts-help: 0.69 (i.e. slower, 23.87s (σ 0.08) to 34.75s (σ 0.07))
schml-specify-rep: 0.94 (i.e. slower, 9.93s (σ 0.03) to 10.59s (σ 0.47))
```