@samth 2017-10-16
What if you just try a simple solution and memoize the computation of recursive-kinds with a hash table?