@bennn 2017-10-17
I'm not sure how to memoize.
At first I tried making a function from `sc` to `recursive-kinds` that used the `sc` as an `eq?` cache key, but that's a bad key because the optimizer changes the `sc`.

The commit I just pushed here adds a new function `instantiate/optimize` that computes `recursive-kinds` once. It's about 4s slower than master. Looking into improving it.