@bennn 2017-10-17
The latest commit has better performance, 

benchmark               | master | e1a2a77
---------|---|---
acquire                 | 18.19s (σ 0.42) | 20.35s (σ 0.66) 
dungeon                 | 10.08s (σ 0.08) | 19.66s (σ 17.65) 
forth                   | 5.6s (σ 0.05)   | 5.96s (σ 0.13)
fsm                     | 5.65s (σ 0.02)  | 6.06s (σ 0.13)
math-flonum             | 4.53s (σ 0.02)  | 5.03s (σ 0.25)
new-metrics             | 3.79s (σ 0.1)   | 4.02s (σ 0.2)
old-metrics             | 2.87s (σ 0.04)  | 2.95s (σ 0.04)
parser                  | 2.43s (σ 0.02)  | 2.46s (σ 0.04)
schml-interp-casts-help | 23.87s (σ 0.08) | 25.77s (σ 0.48)
schml-specify-rep       | 9.93s (σ 0.03)  | 10.61s (σ 0.18)