@samth 2017-10-17
What if we have a mutable field on all `sc`s, which we update with the kind when we learn it, and transfer when we optimize? That should fix the problem of keys changing.