@bennn 2017-10-18
I tried a simple version of that for `name/sc` here: https://github.com/racket/typed-racket/pull/631/commits/fc15febb5e170be9ea8a2c93ef3cbc6c53c5dce2
but it seemed slow after 1 measurement so I moved on to the `free-id-set` version in this commit.

I'll try it again.