@bennn 2017-10-19
This PR currently: (1) solves for the kinds of recursive contracts before optimizing and (2) shares the result with `optimize` and `instantiate`. Performance seem ok (see below). I think this is the right fix.

I thought about adding a mutable "kind" field to `name/sc` and `recursive-sc`, but two problems stopped me:

1. (hard) The constraint solver works with identifiers, not scs. I didn't see an easy way to jump from an identifier back to its parent `name/sc` or `recursive-sc`, and didn't want to traverse the whole `sc` again to update those fields with data from a `recursive-kinds` hashtable.
2. (more refactoring) If `sc`s know their kind, then `instantiate` probably doesn't need the table of recursive kinds at all, and should be refactored

- - -

Here's a table with compile times from the `tr-performance` repo.
- `master` is TR commit 26f65f1
- `e1a2a77` adds extra state to the implementation of `name/sc`
- `f59df64` has `optimize` and `instantiate` share the table of `recursive-kinds`
- `e7be251` is like `f59df64`, but with those `make-XXX` functions in `optimize.rkt` to save a little performance

   benchmark               | master          | e1a2a77          | f59df64          | e7be251
  -------------------------|-----------------|------------------|------------------|------------
   acquire                 | 18.19s (σ 0.42) | 20.35s (σ 0.66)  | 19.91s (σ 1.02)  | 19.86s (σ 1.37)
   dungeon                 | 10.08s (σ 0.08) | 19.66s (σ 17.65) | 25.67s (σ 20.22) | 21.51s (σ 18.01)
   forth                   | 5.6s (σ 0.05)   | 5.96s (σ 0.13)   | 5.81s (σ 0.03)   | 5.75s (σ 0.04)
   fsm                     | 5.65s (σ 0.02)  | 6.06s (σ 0.13)   | 6.05s (σ 0.07)   | 5.98s (σ 0.02)
   math-flonum             | 4.53s (σ 0.02)  | 5.03s (σ 0.25)   | 6.44s (σ 0.99)   | 4.85s (σ 0.02)
   new-metrics             | 3.79s (σ 0.1)   | 4.02s (σ 0.2)    | 4.79s (σ 0.64)   | 4.04s (σ 0.1)
   old-metrics             | 2.87s (σ 0.04)  | 2.95s (σ 0.04)   | 3.28s (σ 0.21)   | 3.06s (σ 0.02)
   parser                  | 2.43s (σ 0.02)  | 2.46s (σ 0.04)   | 2.8s (σ 0.21)    | 2.62s (σ 0.03)
   schml-interp-casts-help | 23.87s (σ 0.08) | 25.77s (σ 0.48)  | 26.0s (σ 0.12))  | 24.41s (σ 0.05)
   schml-specify-rep       | 9.93s (σ 0.03)  | 10.61s (σ 0.18)  | 11.16s (σ 0.42)  | 10.64s (σ 0.09)