@samth 2017-10-19
What is going on with `dungeon`?

Also, can you run this with the `tr-timing` logging, and see where the extra time is? In particular, is the extra time actually in this solving, or in the expansion of the (now-correct) contracts?