@pnwamk 2017-10-19
Speculation: the Schml compiler hits awful non-deterministic performance in our contract generation -- I believe caused by the iteration order of free-id-tables/sets.

This could be why the σ is so high for dungeon.

If so, I think we should experiment with (1) using a deterministic dictionary for free ids, or (2) see if the cost of sorting the free ids is prohibitive (I worry it will be, but we should experiment and see).