@pnwamk 2017-10-19
I phrased it poorly -- I wonder if the same issue Schml is hitting in contract generation, Dungeon is also hitting. With the σ on the measurements that high, that was just my first thought of another program high highly variable contract generation times under TR.