@pnwamk 2017-10-19
_meta comment -- I haven't examined in great detail the most recent changes, so it may not even totally apply, and I'm not suggesting Ben should go try this now:_ I wonder if we just refactored our static-contract representation to keep around more relevant/useful data if this all wouldn't just be simpler to do correctly (i.e. at a static-contract's definition site, we specify how its kind is calculated and cached, etc (similar to our DSL for Reps (see `rep/rep-utils.rkt`))