@samth 2017-10-19
That might be helpful as well with #633, where the issue I'm running into is wanting to go from an sc to a type so that I know what type a contract came from.