@bennn 2017-10-19
Part of the reason I don't want to do the refactoring here: IMO this PR is an "emergency fix" because programs that used to run now error (after https://github.com/racket/typed-racket/commit/9df037b0f6e6b2df3cc0993631bb2870367e291a). 

[[ I was really hoping to fix #628 with this before the release (just like #625) ]]